% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssr_roclet.R
\name{roclet_output.roclet_rssr}
\alias{roclet_output.roclet_rssr}
\title{roclet_output.roclet_rssr}
\usage{
roclet_output.roclet_rssr(x, results, base_path)
}
\arguments{
\item{x}{A \code{roclet} object.}

\item{results}{Value returned from your \code{roclet_process()} method.}

\item{base_path}{Path to root of source package.}
}
\description{
roclet_output.roclet_rssr
}
