% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rssr_roclet.R
\name{roclet_process.roclet_rssr}
\alias{roclet_process.roclet_rssr}
\title{roclet_process.roclet_rssr}
\usage{
roclet_process.roclet_rssr(x, blocks, env, base_path)
}
\arguments{
\item{x}{A \code{roclet} object.}

\item{blocks}{A list of \link[roxygen2]{roxy_block} objects.}

\item{env}{Package environment.}

\item{base_path}{Path to root of source package.}
}
\description{
roclet_process.roclet_rssr
}
